requirements.txt
numpy
matplotlib
flask
Pillow